chore(graph): rename kb/mesh.py to kb/vault_activity.py
kb/mesh.py and kb/knowledge_mesh.py competed for the name "mesh":
mesh.py held MeshState, the live Vault Activity projection (per
CONTEXT.md), while knowledge_mesh.py held the actual Knowledge Mesh.
The collision left two modules with one name and no module docstring
on the older file.

Rename kb/mesh.py to kb/vault_activity.py to match CONTEXT.md
vocabulary, leaving knowledge_mesh.py as the unambiguous owner of
"mesh". Add a module docstring to vault_activity.py stating it
implements Vault Activity, and fix knowledge_mesh.py's docstring
reference to the old kb.mesh.MeshState path. Update every importer in
kb/, tests/, and scripts/. Route paths (/api/mesh, /api/mesh/graph)
are unchanged, as requested by the issue.

Full test suite (1727 passed, 1 skipped) and ruff check . remain
green.

Fixes #121
