+++ kb/self_improve.py
@@ -26,7 +26,7 @@ from kb.llm_enrichment import (
     parse_json_payload,
     redacted_preview,
 )
-from kb.mesh import MeshState
+from kb.vault_activity import MeshState
 from kb.service import Citadel
 
 logger = logging.getLogger(__name__)
